requirements.txt
Django==2.1.4
django-filter==2.0.0
django-money==0.14.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
psycopg2==2.7.6.1
py-moneyed==0.8.0
PyJWT==1.7.1
pytz==2018.7
six==1.12.0
drf-aggregates==0.0.10
pandas==0.22.0